VS 2013 project folder + files
I copied the VB files currently on github into the project folder
Valentin added to the Google drive. This entire folder is the Visual
Basic project in Visual Studio 2013. If you run VS 2013 and use open
project and navigate to the github folder and select the Studio Solution
file to run.